pyproject.toml
[project]
name = "he-kit"
version = "0.0.16"
description = "A project generator that bootstraps a FastAPI setup following Helicon's standards and best practices."
authors = [{ name = "gs", email = "gs@helicon.ai" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.0",
    "fastapi[standard]>=0.119.1",
    "pydantic-settings>=2.11.0",
    "sqlmodel>=0.0.27",
    "sqlalchemy>=2.0.44",
    "structlog>=25.4.0",
    "python-jose[cryptography]>=3.5.0",
]

[project.optional-dependencies]
keycloak = [
    "python-keycloak>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/he_kit"]

[dependency-groups]
dev = [
    "jinja2>=3.1.6",
    "pytest>=8.4.2",
    "pre-commit>=4.3.0",
    "pytest-asyncio>=1.2.0",
    "testcontainers>=4.13.2",
    "typer>=0.19.2",
    "python-keycloak>=4.6.0",
]

[project.scripts]
he-kit = "he_kit.cli:app"
